requirements.txt
alipay-sdk-python==3.7.78
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
hiredis==2.3.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
pillow==10.2.0
pyasn1==0.6.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
rpds-py==0.18.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
